GetRecordings/requirements.txt
boto3
mysql-connector-python
pandas
enlighten
uuid
numpy
soundfile